pyproject.toml
[project]
name = "situational-awareness-tracker"
version = "0.1.0"
description = "Weekly watcher for Situational Awareness LP 13F filings on SEC EDGAR"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "lxml>=5.0",
    "yfinance>=0.2.40",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "resend>=2.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25",
    "ruff>=0.5",
]

[project.scripts]
tracker = "run:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 110
target-version = "py311"
